use cbor encoding, allow middleware skip
Updates the cookie serializer to use the [CBOR
codec](https://github.com/fxamacker/cbor) for encoding the cookie data,
as well as adds an option to skip TemplMiddleware for certain paths.

In testing, the CBOR encoding is 3.5x faster than encoding/gob. It also
doesn't require the struct data to be registered for encoding, which is
an additional UX improvement over gob.